feat(*): Adds basic tracing in to exported packages
This adds some basic logging around requests and changes out all `println!`
macros for logging. There are definitely more places we could log (like doing
a `trace!` on memory offsets and read/write operations), but this was a good
start and got everything going.

Right now, `tracing` is set up with the `log` feature, which means if a consumer
of the library doesn't have a subscriber set up, but _is_ using `log`, it will
still function like a normal logger. Otherwise, it will use the configured subscriber
provided by the user