Fix header parsing in AssemblyScript
This commit fixes the way headers are parsed in AssemblyScript.
Specifically, after the request headers were not sent as JSON anymore,
there was an additional pair of quotation marks for both header key and
value that wasn't properly escaped.

This change removes the additional quotation pair and brings the
AssemblyScript way of passing strings back in sync with the way this is
done in Rust.

Signed-off-by: Radu M <root@radu.sh>