Check response header length in tests
This commit updates the Rust and AssemblyScript tests to check the
length of the response headers and error out if they change.

This ensures we periodically test for any changes in the way headers are
parsed, but has the disadvantage that if any of the servers we are using
to test changes their response headers, we need to update the tests.

Signed-off-by: Radu M <root@radu.sh>