Disable `allowed_hosts` by default
This commit renames `allowed_domains` to `allowed_hosts` and defaults
to NOT allowing guest modules to make HTTP requests when `None` is
the value passed.

This change brings the way this library handles its "capability" closer
to how upstream WASI works, while also making more sense (if `None` is
passed as value for `allowed_hosts`, modules are not allowed to send
requests).

Signed-off-by: Radu M <root@radu.sh>