Change the API to allow streaming & no callback to alloc
This makes a couple changes to the API to make it closer to other
WASI APIs.

`req()` returns a handle to a `Response` object, along with the status
code.

A `Response` object can be then used to:
- Read individual headers
- Read the body in a streaming fashion.

A `body_read()` function pulls as many bytes as the client wants,
and returns `0` when the end of the stream is reached.
This makes it compatible with streaming, doesn't require any dynamic
memory allocation, and allows the client to control the maximum
response size it's willing to accept.

For conveniency, the Rust and AssemblyScript bindings also have a
`readAll` function that reads everything until the end and puts that
into a dynamically allocated array.

Usage in an app using wasmtime is now similar to wasi (core) and other
WASI proposals, with an object and an `add_to_linker()` function.

Errors are a little bit easier to keep track of, by using Rust errors,
that are converted to return codes when needed.
WITX is supposed to provide eventually provide the magic to get error
strings from that.

Dependencies have been updated, especially `wasmtime`.