Add option to limit maximum concurrent requests
This commit adds a runtime option to limit the maximum number of
concurrent requests a guest module can make.
If set, the runtime implementation checks the current number of open
handles and if it exceeds the limit, the current request will fail.

Note that this does not affect the default (set as `None`), in which
case guests can send as many requests as they want.

This affects the "un-closed" handles - for example, if a Rust `Response`
goes out of scope, `Drop` will be automatically executed and the handle
closed, so the request does not count towards the limit.
On the other hand, the current AssemblyScript implementation requires
the guest to manually close the handle, so this could have an impact.
A subsequent change to the AssemblyScript implementation should
implement a GC finalizer for `Response`.

Signed-off-by: Radu M <root@radu.sh>