Remove client dep to fix Windows
This commit removes the dependency on the client library from the
Wasmtime library in order to fix a Windows linking error.
Specifically, the Windows error stems from the fact that the client
library, which is mainly used for building Wasm modules, imports the
functions related to sending HTTP requests from the runtime.
There seems to be a bug in the Windows linker resulting in a linking
error, even if the Wasmtime crate is not calling any of the external
symbols, and this commit removes that dependency entirely, given that
the used functionality was limited to a single function.
See https://github.com/rust-lang/rust/issues/86125.

Signed-off-by: Radu M <root@radu.sh>