feat: allow passing context data per instance
Previously, context data (allowed hosts and maximum number of concurrent
requests) was passed when linking the functionality, and couldn't be
controlled if the same linker instance was to instantiate multiple
modules with different policies.

This makes it possible to split that and pass context data when
creating the store, rather than when creating the linker.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>