refactor!: change type of `HttpState::add_to_linker`'s `get_cx` parameter
BREAKING CHANGE: The type of this parameter has changed from `impl Fn(&T) ->
&HttpCtx + Send + Sync + Copy + 'static` to `impl Fn(&T) -> HttpCtx + Send +
Sync + 'static`.  Besides removing the `Copy` bound (which wasn't needed), I've
changed the return type from `&HttpCtx` to `HttpCtx`.

The rationale for this is that the return value of `get_cx` was effectively
being cloned anyway immediately after calling it, so there was no efficiency
gain from returning a reference.  Also, any implementation of the original type
would have to ensure the returned reference has the same lifetime as the context
parameter, which essentially means the implementation is forced to store the
`HttpCtx` inside the context.  That's overkill for e.g. `wagi`, which has no
need for per-context HTTP configuration.

Signed-off-by: Joel Dice <joel.dice@gmail.com>